Factor out common code from FileExtentlist and FileBtree
