+++ src/libxfuse/file_extent_list.rs
@@ -25,16 +25,14 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-use std::{
-    cmp::min,
-    io::{BufRead, Seek, SeekFrom},
-};
+use std::io::{BufRead, Seek};
+
+use bincode::de::read::Reader;
 
 use super::{
     bmbt_rec::BmbtRec,
     definitions::{XfsFileoff, XfsFsblock, XfsFsize},
     file::File,
-    sb::Sb,
 };
 
 #[derive(Debug)]
@@ -44,10 +42,14 @@ pub struct FileExtentList {
     pub block_size: u32,
 }
 
-impl FileExtentList {
+impl<R: BufRead + Reader + Seek> File<R> for FileExtentList {
+    fn block_size(&self) -> u32 {
+        self.block_size
+    }
+
     /// Return the extent, if any, that contains the given data block within the file.
     /// Return its starting position as an FSblock, and its length in file system block units
-    pub fn get_extent(&self, block: XfsFileoff) -> (Option<XfsFsblock>, u64) {
+    fn get_extent(&self, _buf_reader: &mut R, block: XfsFileoff) -> (Option<XfsFsblock>, u64) {
         match self.bmx.partition_point(|entry| entry.br_startoff <= block) {
             0 => {
                 // A hole at the beginning of the file
@@ -72,35 +74,8 @@ impl FileExtentList {
             }
         }
     }
-}
 
-impl<R: BufRead + Seek> File<R> for FileExtentList {
-    fn read(&mut self, buf_reader: &mut R, _super_block: &Sb, offset: i64, size: u32) -> Vec<u8> {
-        let mut remaining_size = min(size as i64, self.size - offset);
-        let mut data = Vec::<u8>::with_capacity(remaining_size.max(0) as usize);
-
-        let mut logical_block = offset / i64::from(self.block_size);
-        let mut block_offset = offset % i64::from(self.block_size);
-
-        while remaining_size > 0 {
-            let (blk, blocks) = self.get_extent(logical_block as u64);
-            let z = min(remaining_size, (blocks as i64 * self.block_size as i64) - block_offset);
-            let oldlen = data.len();
-            data.resize(oldlen + z as usize, 0u8);
-            if let Some(blk) = blk {
-                buf_reader
-                    .seek(SeekFrom::Start(blk * u64::from(self.block_size) + block_offset as u64))
-                    .unwrap();
-
-                buf_reader.read_exact(&mut data[oldlen..]).unwrap();
-            } else {
-                // A hole
-            }
-            logical_block += blocks as i64;
-            remaining_size -= z;
-            block_offset = 0;
-        }
-
-        data
+    fn size(&self) -> XfsFsize {
+        self.size
     }
 }
